requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
geographiclib==2.1
geopy==2.4.1
gunicorn==21.2.0
dj-database-url==2.3.0
Pillow>=9.0.0
psycopg2-binary==2.9.11
PyJWT==2.8.0
python-dotenv==1.2.1
reportlab>=4.0
openpyxl>=3.1.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.8.2
